refactor: 음식점 리뷰와 음식점 리뷰 좋아요 인수테스트 합침 (#583)
